test/02_deploy_execute_erc20.spec.js
