+++ test/testList.template.txt
@@ -2,7 +2,7 @@
 02_deploy_execute_flipcoin.spec.js
 02_deploy_execute_millionaire.spec.js
 02_deploy_execute_voting.spec.js
-02_deploy_erc20.spec.js
+02_deploy_execute_erc20.spec.js
 02_deploy_execute_constructor_set_state_12.spec.js
 03_deploy_fail_bytecode.spec.js
 03_deploy_fail_outofgas.spec.js
@@ -10,7 +10,6 @@
 03_deploy_fail_wrong_encryption_key.spec.js
 03_deploy_fail_wrong_eth_address.spec.js
 03_deploy_fail_wrong_worker.spec.js
-10_execute_erc20.spec.js
 20_execute_fail_outofgas.spec.js
 20_execute_fail_parameters.spec.js
 20_execute_fail_wrong_encryption_key.spec.js
